// src/enums/log.enum.ts
export enum LogEventType {
  BOOK_SERVICE_CLICK = "BOOK_SERVICE_CLICK",
  BOOK_SERVICE_CONFIRM = "BOOK_SERVICE_CONFIRM",
  BOOKING_SERVICE_BLOCK = "BOOKING_SERVICE_BLOCK",
  PAYMENT_INITIATED = "PAYMENT_INITIATED",
  PAYMENT_SUCCESS = "PAYMENT_SUCCESS",
  PAYMENT_FAILED = "PAYMENT_FAILED",
  PAYMENT_REFUNDED = "PAYMENT_REFUNDED",
  PAYMENT_REFUNDED_FAILED = "PAYMENT_REFUNDED_FAILED",
  SERVICE_PROVIDER_CHANGED = "SERVICE_PROVIDER_CHANGED",
  SERVICE_PROVIDER_ASSIGNED = "SERVICE_PROVIDER_ASSIGNED",
  SERVICE_COMPLETED = "SERVICE_COMPLETED",
  SERVICE_STATUS_CHANGED = "SERVICE_STATUS_CHANGED",
  SERVICE_CANCELLED = "SERVICE_CANCELLED",
  BOOKING_STATUS_CHANGED = "BOOKING_STATUS_CHANGED",
}

export enum LogCategory {
  PAYMENT = "PAYMENT",
  BOOKING = "BOOKING",
  SERVICE = "SERVICE",
  SERVICE_PROVIDER = "SERVICE_PROVIDER",
  CUSTOMER = "CUSTOMER",
}

export enum LogStatus {
  INITIATED = "INITIATED",
  SUCCESS = "SUCCESS",
  FAILED = "FAILED",
}

export enum CurrencySymbol {
  USD = "$",
  INR = "₹",
}

export enum CurrencyValueSymbol {
  USD = "USD",
  INR = "INR",
}